+++ src/parse-query/parse-combo-data.ts
@@ -1,10 +1,12 @@
 import type { SearchParameters } from "../types";
 
-const COMBO_DATA_TYPES: ["prerequisites", "steps", "results"] = [
+export const COMBO_DATA_TYPES: [
+  "cards",
   "prerequisites",
   "steps",
-  "results",
-];
+  "results"
+] = ["cards", "prerequisites", "steps", "results"];
+const SUPPORTED_OPERATORS_FOR_DATA = [":", "="];
 
 export default function parseComboData(
   params: SearchParameters,
@@ -17,6 +19,35 @@ export default function parseComboData(
 
   COMBO_DATA_TYPES.forEach((dataType) => {
     if (dataType.indexOf(normalizedKey) === -1) {
+      return;
+    }
+
+    if (operator !== ":" && Number(value) >= 0) {
+      if (isNegativeKey) {
+        params.errors.push({
+          key,
+          value,
+          message: `The key "${key}" does not support operator "${operator}"`,
+        });
+
+        return;
+      }
+
+      params[dataType].sizeFilters.push({
+        method: operator,
+        value: Number(value),
+      });
+
+      return;
+    }
+
+    if (!SUPPORTED_OPERATORS_FOR_DATA.includes(operator)) {
+      params.errors.push({
+        key,
+        value,
+        message: `Operator ${operator} is not compatible with key "${key}" and value "${value}"`,
+      });
+
       return;
     }