+++ test/unit/parse-query/index.test.ts
@@ -1,9 +1,7 @@
 import parseQuery from "../../../src/parse-query";
-import parseCardQuery from "../../../src/parse-query/parse-card-query";
 import parseColorIdentity from "../../../src/parse-query/parse-color-identity";
 import parseComboData from "../../../src/parse-query/parse-combo-data";
 
-jest.mock("../../../src/parse-query/parse-card-query");
 jest.mock("../../../src/parse-query/parse-color-identity");
 jest.mock("../../../src/parse-query/parse-combo-data");
 
@@ -11,20 +9,20 @@ describe("parseQuery", () => {
   it("parses plain text into cards", () => {
     parseQuery("foo bar baz");
 
-    expect(parseCardQuery).toBeCalledTimes(3);
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledTimes(3);
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "foo"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "bar"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
@@ -36,7 +34,7 @@ describe("parseQuery", () => {
   it("supports an empty string for query", () => {
     const result = parseQuery("");
 
-    expect(parseCardQuery).not.toBeCalled();
+    expect(parseComboData).not.toBeCalled();
     expect(parseColorIdentity).not.toBeCalled();
     expect(parseComboData).not.toBeCalled();
 
@@ -75,20 +73,20 @@ describe("parseQuery", () => {
       "foo unknown:value unknown2:'value 2' unknown3:\"value 3\" bar -card>3"
     );
 
-    expect(parseCardQuery).toBeCalledTimes(3);
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledTimes(3);
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "foo"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "bar"
     );
-    expect(parseCardQuery).toBeCalledWith(expect.anything(), "-card", ">", "3");
+    expect(parseComboData).toBeCalledWith(expect.anything(), "-card", ">", "3");
 
     expect(result).toEqual(
       expect.objectContaining({
@@ -118,39 +116,37 @@ describe("parseQuery", () => {
       "Kiki ci:wbr -ci=br card:Daxos card:'Grave Titan' card:\"Akroma\" unknown:value -card:Food prerequisites:prereq steps:step results:result -prerequisites:xprereq -steps:xstep -result:xresult"
     );
 
-    expect(parseCardQuery).toBeCalledTimes(5);
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledTimes(11);
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "Kiki"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "Daxos"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "Grave Titan"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "Akroma"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "-card",
       ":",
       "Food"
     );
-
-    expect(parseComboData).toBeCalledTimes(6);
     expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "prerequisites",
@@ -220,39 +216,37 @@ describe("parseQuery", () => {
       "Kiki CI:wbr CARD:Daxos CARD:'Grave Titan' CARD:\"Akroma\" UNKNOWN:value -CARD:Food PREREQUISITES:prereq STEPS:step RESULTS:result -PREREQUISITES:xprereq -STEPS:xstep -RESULT:xresult"
     );
 
-    expect(parseCardQuery).toBeCalledTimes(5);
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledTimes(11);
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "Kiki"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "Daxos"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "Grave Titan"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "Akroma"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "-card",
       ":",
       "Food"
     );
-
-    expect(parseComboData).toBeCalledTimes(6);
     expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "prerequisites",
@@ -316,39 +310,37 @@ describe("parseQuery", () => {
       "Kiki c_i:wbr c_ar_d:Daxos ca_rd:'Grave Titan' ca_rd:\"Akroma\" unknow_n:value -c_ard:Food _prere_quisit_es_:prereq st_eps:step r_esu_lts:result -prer_equisites:xprereq -ste_ps:xstep -res_ult:xresult"
     );
 
-    expect(parseCardQuery).toBeCalledTimes(5);
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledTimes(11);
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "Kiki"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "Daxos"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "Grave Titan"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
       "Akroma"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "-card",
       ":",
       "Food"
     );
-
-    expect(parseComboData).toBeCalledTimes(6);
     expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "prerequisites",
@@ -433,20 +425,20 @@ describe("parseQuery", () => {
       `${kind}:Rashmi ${kind}:"with quotes", -${kind}:'excluded thing'`
     );
 
-    expect(parseCardQuery).toBeCalledTimes(3);
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledTimes(3);
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       kind,
       ":",
       "Rashmi"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       kind,
       ":",
       "with quotes"
     );
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       `-${kind}`,
       ":",
@@ -457,8 +449,8 @@ describe("parseQuery", () => {
   it("can parse names with apostrophes", () => {
     parseQuery(`card:"Freyalise, Llanowar's Fury"`);
 
-    expect(parseCardQuery).toBeCalledTimes(1);
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledTimes(1);
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",
@@ -469,8 +461,8 @@ describe("parseQuery", () => {
   it("can parse names with quotes in the name", () => {
     parseQuery(`card:'Kongming, "Sleeping Dragon"'`);
 
-    expect(parseCardQuery).toBeCalledTimes(1);
-    expect(parseCardQuery).toBeCalledWith(
+    expect(parseComboData).toBeCalledTimes(1);
+    expect(parseComboData).toBeCalledWith(
       expect.anything(),
       "card",
       ":",